requirements.txt
torch
torchvision
pillow
accelerate
transformers